requirements.txt
openai-whisper
watchdog
openai
python-dotenv